Many changes to module system
You now need to run make generate to generate linkdefs and the undefines
for the linker script.
LinkDefs are definitions for all available and exported symbols for the
current elf. You may need to build a total of two times (make build;
make generate; make build) in order to have a complete list of available
symbols ready.
The modules now use C++ instead of C, and they use the includes from
WiMu's source.
Some fixes to how R_PPC_ADDR16 is relocated, for the better.