Module System updates
Module Makefiles now link to libstdc++, since internal __cxa calls are
internal to the ELFs themselves, and cannot be linked from WiMu to the
modules.

The test module now adds a RenderableText to the screen to demonstrate
it's ability to create objects and call methods from WiMu.

Main now doesn't call the module if linking fails.

Updated generated files.